IMP - Refactored Google Sign-in
